api/get_popularity/requirements.txt
google-cloud-bigquery==3.*
functions-framework==3.*